requirement.txt
mcp>=0.1.0
google-generativeai
python-dotenv
yfinance
requests
beautifulsoup4
httpx